fix
fix

+++ Assets/Script/CardEffectAndCulDuringManager.cs
@@ -213,7 +213,7 @@ public class CardEffectAndCulDuringManager : MonoBehaviour
     public void RequsetDownUserCurrentStatFromDamage(UserAoubtDamage givenDamge, bool isAdd)
     {
         if(isAdd)
-            userAboutDamages[givenDamge.targetMonsterID].damage.Add(givenDamge);
+            userAboutDamages[(int)givenDamge.targetMonsterID].damage.Add(givenDamge);
         int targetUserID = (int)givenDamge.targetMonsterID;
         float totalDamage = givenDamge.hitHpDamage + givenDamge.hitDGDamage;
         userTotalStates[targetUserID].currentHp -= totalDamage * userTotalStates[targetUserID].damageMultiplier;
